requirements.txt
flask
flask-cors
yt-dlp
spotifyscraper
pyinstaller
websocket-client
requests
certifi
Flask-Limiter
setuptools-rust
